pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "company-rag"
version = "0.1.0"
description = "RAG knowledge base for US NYSE publicly traded companies"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "llama-index-core>=0.10",
    "llama-index-readers-file>=0.1",
    "sentence-transformers>=2.2",
    "chromadb>=0.4",
    "openai>=1.0",
    "sec-edgar-downloader>=5.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
